requirements.txt
beautifulsoup4==4.12.3
lxml==5.2.2
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.37.1
pydeck==0.9.1
